Rename utils files
stack-info: PR: https://github.com/QSTheory/fftarray/pull/165, branch: SJSeckmeyer/stack/18
