+++ fftarray/_utils/formatting.py
